Consolidate some RhiSystem tests
Remove VulkanIntegration tests, it was a bit of a pointless split. Move some of the tests to TestRhiSystem, delete the rest for not being useful
