+++ tests/unit_tests/rhi_system/TestRhiSystem.cpp
@@ -25,22 +25,27 @@ protected:
         }
     }
 
+    rhi::RhiSystemCreateInfo GetBaseCreateInfo()
+    {
+        rhi::RhiSystemCreateInfo createInfo{};
+        createInfo.ApplicationName = "RhiSystemTest";
+        createInfo.EngineName = "DiamondDogsTestEngine";
+        createInfo.AppVersion = VK_MAKE_VERSION(1, 0, 0);
+        createInfo.EngineVersion = VK_MAKE_VERSION(0, 1, 0);
+        createInfo.VkVersion = rhi::ApiVersion::Vulkan13;
+        createInfo.ValidationLevel = rhi::ValidationLayers::BaseOnly;
+        createInfo.RequiredInstanceExtensions = {"VK_EXT_debug_utils"};
+        createInfo.ShaderCacheDir = tempShaderCacheDir.string();
+        return createInfo;
+    }
+
     std::filesystem::path tempShaderCacheDir;
 };
 
 // Test basic RHI system construction with CreateInfo
 TEST_F(RhiSystemTest, BasicCreateInfoConstruction)
 {
-    rhi::RhiSystemCreateInfo createInfo{};
-    createInfo.ApplicationName = "RhiSystemTest";
-    createInfo.EngineName = "DiamondDogsTestEngine";
-    createInfo.AppVersion = VK_MAKE_VERSION(1, 0, 0);
-    createInfo.EngineVersion = VK_MAKE_VERSION(0, 1, 0);
-    createInfo.VkVersion = rhi::ApiVersion::Vulkan13;
-    createInfo.ValidationLevel = rhi::ValidationLayers::BaseOnly;
-    createInfo.RequiredInstanceExtensions = {"VK_EXT_debug_utils"};
-    createInfo.ShaderCacheDir = tempShaderCacheDir.string();
-
+    auto createInfo = GetBaseCreateInfo();
     EXPECT_NO_THROW({
         auto rhiSystem = std::make_unique<rhi::RhiSystem>(createInfo);
         EXPECT_NE(rhiSystem, nullptr);
@@ -60,6 +65,66 @@ TEST_F(RhiSystemTest, MinimalConfiguration)
         auto rhiSystem = std::make_unique<rhi::RhiSystem>(createInfo);
         EXPECT_NE(rhiSystem, nullptr);
     });
+}
+
+// Test Vulkan 1.0 API version, to see if we can still create an instance of this version
+TEST_F(RhiSystemTest, Vulkan10ApiVersion) {
+    auto createInfo = GetBaseCreateInfo();
+    createInfo.VkVersion = rhi::ApiVersion::Vulkan10;
+
+    EXPECT_NO_THROW({
+        auto rhiSystem = std::make_unique<rhi::RhiSystem>(createInfo);
+        EXPECT_NE(rhiSystem, nullptr);
+        EXPECT_NE(rhiSystem->GetInstance(), nullptr);
+    });
+}
+
+// Test Vulkan 1.1 API version (another key version for use)
+TEST_F(RhiSystemTest, Vulkan11ApiVersion) {
+    auto createInfo = GetBaseCreateInfo();
+    createInfo.VkVersion = rhi::ApiVersion::Vulkan11;
+
+    EXPECT_NO_THROW({
+        auto rhiSystem = std::make_unique<rhi::RhiSystem>(createInfo);
+        EXPECT_NE(rhiSystem, nullptr);
+        EXPECT_NE(rhiSystem->GetInstance(), nullptr);
+    });
+}
+
+// Test Latest API version
+TEST_F(RhiSystemTest, LatestApiVersion) {
+    auto createInfo = GetBaseCreateInfo();
+    createInfo.VkVersion = rhi::ApiVersion::Latest;
+
+    EXPECT_NO_THROW({
+        auto rhiSystem = std::make_unique<rhi::RhiSystem>(createInfo);
+        EXPECT_NE(rhiSystem, nullptr);
+        EXPECT_NE(rhiSystem->GetInstance(), nullptr);
+    });
+}
+
+// Verify that we can create an instance with different validation layers
+TEST_F(RhiSystemTest, ValidationLayers)
+{
+    std::vector<rhi::ValidationLayers> validationLevels
+    {
+        rhi::ValidationLayers::None,
+        rhi::ValidationLayers::BaseOnly,
+        rhi::ValidationLayers::WithSynchronization,
+        rhi::ValidationLayers::Full
+    };
+
+    for (const auto& level : validationLevels)
+    {
+        auto createInfo = GetBaseCreateInfo();
+        createInfo.ValidationLevel = level;
+
+        EXPECT_NO_THROW({
+            auto rhiSystem = std::make_unique<rhi::RhiSystem>(createInfo);
+            EXPECT_NE(rhiSystem, nullptr);
+            EXPECT_NE(rhiSystem->GetInstance(), nullptr);
+        }) << "Failed for validation level: " << static_cast<int>(level);
+    }
 }
 
 // Test RHI system with presentation support
